check_sorted + special case_3 numbers
Push_swap now behaves correctly when the stack is already sorted;
Push_swap now sorts the array in the most efficient possible way when the array contains 3 numbers. It does so by comparing the parameters that the program receives before even creating a stack. Most likely will have to change it to sort only after the stack was created, so i can use the special case_3 sorter when sorting an array of 4 and 5 integers.